FuentesCurso/UD 06. Docker Compose/Ficheros casos practicos UD06/CasoPractico5-WhisperAI-TelegramBot/telegram-bot-whisper-cpp-main/requirements.txt
numpy
python_telegram_bot
python-telegram-bot[all]
whispercpp
